Refactor common functions to utils library
